move container to the containers folder
